+++ segment_anything/README.md
@@ -12,10 +12,10 @@ pip install -r requirements.txt
 Download checkpoints from [Segment Anything repo](https://github.com/facebookresearch/segment-anything/tree/main?tab=readme-ov-file#model-checkpoints) and put them under `models/pt_models`
 
 ```bash
-python convert_pt.py --pt-path=models/pt_models/[sam_vit_b_01ec64.pth | sam_vit_h_4b8939.pth | sam_vit_l_0b3195.pth]
+python scripts/convert_pt.py --pt-path=models/pt_models/[sam_vit_b_01ec64.pth | sam_vit_h_4b8939.pth | sam_vit_l_0b3195.pth]
 ```
 The `safetensors` weights are generated under `models/mlx_models`
 
 ## Get Started
 
-Take `predictor_example.ipybn` as an example to try Segment Anything Model with MLX.
+See examples `notebooks/predictor_example.ipynb` and `notebooks/automatic_mask_generator_example.ipynb` to try Segment Anything Model with MLX.
