+++ segment_anything/segment_anything/__init__.py
@@ -1,0 +1,9 @@
+from .automatic_mask_generator import SamAutomaticMaskGenerator
+from .build_sam import (
+    build_sam,
+    build_sam_vit_h,
+    build_sam_vit_l,
+    build_sam_vit_b,
+    sam_model_registry,
+)
+from .predictor import SamPredictor
